CB#7 - Refactoring - Fix structure in features
