+++ src/crypto-bank.WebAPI/Features/Accounts/Requests/GetNumberOfOpenedAccounts.cs
@@ -1,33 +1,10 @@
-using crypto_bank.Database;
-using JetBrains.Annotations;
 using MediatR;
-using Microsoft.EntityFrameworkCore;
 
 namespace crypto_bank.WebAPI.Features.Accounts.Requests;
 
-public class GetNumberOfOpenedAccounts
+public partial class GetNumberOfOpenedAccounts
 {
     public record Request(DateTimeOffset Begin, DateTimeOffset End) : IRequest<Response>;
 
     public record Response(int NumberOfAccounts);
-
-    [UsedImplicitly]
-    public class RequestHandler : IRequestHandler<Request, Response>
-    {
-        private readonly CryptoBankDbContext _dbContext;
-
-        public RequestHandler(CryptoBankDbContext dbContext)
-        {
-            _dbContext = dbContext;
-        }
-
-        public async Task<Response> Handle(Request request, CancellationToken cancellationToken)
-        {
-            var numberOfAccounts = await _dbContext.Accounts
-                .Where(account => account.OpenedAt >= request.Begin && account.OpenedAt <= request.End)
-                .CountAsync(cancellationToken);
-
-            return new Response(numberOfAccounts);
-        }
-    }
 }
